ref(node): Extract backend logic into eventbuilders
To make the diff easier to read when we remove the backend,
I cherry-picked this patch that extracts the `eventFromException` and
`eventFromMessage` logic from the backend into their own functions in
it's own branch.

This patch updates node to behave similarily to browser. There should be
no changes required.
